perf(opt): :building_construction: restructure core layout
