+++ apps/opt/core/repositories/TripRepository.ts
@@ -1,0 +1,4 @@
+import { Trip } from '../interfaces';
+import { CRUDRepository } from './CRUDRepository';
+
+export interface TripRepository extends CRUDRepository<Trip, string> {}
