+++ apps/opt/integration/actions/TripActions.ts
@@ -1,7 +1,7 @@
 'use server';
 import { cookies } from 'next/headers';
 
-import { TripType } from '@opt/core/trips/types';
+import { Trip } from '@opt/core/interfaces';
 
 import { TripAdapter } from '../adapters';
 import { ActionResponse } from '../types';
@@ -11,10 +11,10 @@ const token = cookies().get('auth_token')?.value;
 // Crear un nuevo trip
 
 export const createTrip = async (
-  trip: TripType,
+  trip: Trip,
 ): Promise<ActionResponse | undefined> => {
   try {
-    const data = await TripAdapter.createTrip(trip, token!);
+    const data = await TripAdapter.create(trip, token!);
     console.log(data);
     return { success: true, data };
   } catch (error) {
@@ -26,10 +26,10 @@ export const createTrip = async (
 // Actualizar un trip existente
 export const updateTrip = async (
   id: string,
-  trip: TripType,
+  trip: Trip,
 ): Promise<ActionResponse | undefined> => {
   try {
-    await TripAdapter.updateTrip(id, trip, token!);
+    await TripAdapter.update(id, trip, token!);
     return { success: true };
   } catch (error) {
     console.error('Failed to update trip:', error);
@@ -42,7 +42,7 @@ export const deleteTrip = async (
   id: string,
 ): Promise<ActionResponse | undefined> => {
   try {
-    await TripAdapter.deleteTrip(id, token!);
+    await TripAdapter.delete(id, token!);
     return { success: true };
   } catch (error) {
     console.error('Failed to delete trip:', error);
