+++ apps/opt/integration/adapters/TripAdapter.ts
@@ -1,32 +1,29 @@
 // infrastructure/repositories/TripAdapter.ts
-import { TripRepository } from '@opt/core/trips/repository';
-import { TripType } from '@opt/core/trips/types';
+
+import { Trip } from '@opt/core/interfaces';
+import { TripRepository } from '@opt/core/repositories';
 import { apiRequest } from '@opt/integration/api';
 
 import { fetcher } from '../swr/config';
 
 export const TripAdapter: TripRepository = {
-  getAllTrips: async (): Promise<TripType[]> => {
+  getAll: async (): Promise<Trip[]> => {
     return fetcher(`/trips`);
   },
 
-  getTripById: async (id: string): Promise<TripType | null> => {
+  getById: async (id: string): Promise<Trip | null> => {
     return fetcher(`/trips/${id}`);
   },
 
-  createTrip: async (trip: TripType, token: string): Promise<void> => {
+  create: async (trip: Trip, token: string): Promise<void> => {
     await apiRequest(`/trips`, 'POST', token, trip);
   },
 
-  updateTrip: async (
-    id: string,
-    trip: TripType,
-    token: string,
-  ): Promise<void> => {
+  update: async (id: string, trip: Trip, token: string): Promise<void> => {
     await apiRequest(`/trips/${id}`, 'PUT', token, trip);
   },
 
-  deleteTrip: async (id: string, token: string): Promise<void> => {
+  delete: async (id: string, token: string): Promise<void> => {
     await apiRequest(`/trips/${id}`, 'DELETE', token);
   },
 };
